+++ spec/ramda/list_spec.rb
@@ -44,15 +44,35 @@ describe Ramda::List do
     end
   end
 
-  xcontext '#chain' do
+  context '#chain' do
     it 'from docs' do
       duplicate = ->(n) { [n, n] }
       expect(r.chain(duplicate, [1, 2, 3])).to eq([1, 1, 2, 2, 3, 3])
     end
 
-    it 'multiple args' do
-      expect(r.chain(R.compose(R.append, R.head)).call([1, 2, 3])).to eq([1, 2, 3, 1])
+    it 'monad with chain' do
+      monad = Class.new do
+        def chain(fn)
+          fn.call(10)
+        end
+      end.new
+
+      expect(r.chain(Ramda.add(5), monad)).to eq(15)
     end
+
+    it 'monad with bind' do
+      monad = Class.new do
+        def bind(fn)
+          fn.call(10)
+        end
+      end.new
+
+      expect(r.chain(Ramda.add(5), monad)).to eq(15)
+    end
+
+    # it 'multiple args' do
+    #   expect(r.chain(R.compose(R.append, R.head)).call([1, 2, 3])).to eq([1, 2, 3, 1])
+    # end
 
     it 'maps a function (a -> [b]) into a (shallow) flat result' do
       times2 = ->(x) { [x * 2] }
@@ -499,7 +519,7 @@ describe Ramda::List do
     end
   end
 
-  xcontext '#unnest' do
+  context '#unnest' do
     it 'from docs' do
       expect(r.unnest([1, [2], [[3]]])).to eq([1, 2, [3]])
       expect(r.unnest([[1, 2], [3, 4], [5, 6]])).to eq([1, 2, 3, 4, 5, 6])